backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
requests==2.32.5
django-cors-headers==4.9.0
pillow==12.0.0
python-dotenv==1.2.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
pytz==2025.2